requirements.txt
tensorflow==2.15.0
tf-keras==2.15.0
transformers==4.37.0
fasttext
scikit-learn
spacy==3.7.5